+++ deploy.psdeploy.ps1
@@ -3,26 +3,25 @@
 
 # ASSUMPTIONS:
 
- # folder structure of:
- # - RepoFolder
- #   - This PSDeploy file
- #   - ModuleName
- #     - ModuleName.psd1
+# folder structure of:
+# - RepoFolder
+#   - This PSDeploy file
+#   - ModuleName
+#     - ModuleName.psd1
 
- # Nuget key in $ENV:NugetApiKey
+# Nuget key in $ENV:NugetApiKey
 
- # Set-BuildEnvironment from BuildHelpers module has populated ENV:BHProjectName
+# Set-BuildEnvironment from BuildHelpers module has populated ENV:BHProjectName
 
 # Publish to gallery with a few restrictions
-if(
+if (
     $env:BHProjectName -and $env:BHProjectName.Count -eq 1 -and
     $env:BHBuildSystem -ne 'Unknown' -and
     $env:BHBranchName -eq "master" -and
     $env:BHCommitMessage -match '!deploy' -and
     $env:APPVEYOR_BUILD_WORKER_IMAGE -like '*2017*' -and
     $env:APPVEYOR_PULL_REQUEST_NUMBER -eq $null 
-)
-{
+) {
     Deploy Module {
         By PSGalleryModule {
             FromSource $ENV:BHProjectName
@@ -33,8 +32,7 @@ if(
         }
     }
 }
-else
-{
+else {
     "Skipping deployment: To deploy, ensure that...`n" +
     "`t* You are in a known build system (Current: $ENV:BHBuildSystem)`n" +
     "`t* You are committing to the master branch (Current: $ENV:BHBranchName) `n" +
@@ -45,11 +43,10 @@ else
 }
 
 # Publish to AppVeyor if we're in AppVeyor
-if(
+if (
     $env:BHProjectName -and $ENV:BHProjectName.Count -eq 1 -and
     $env:BHBuildSystem -eq 'AppVeyor'
-   )
-{
+) {
     Deploy DeveloperBuild {
         By AppVeyorModule {
             FromSource $ENV:BHProjectName
@@ -59,4 +56,5 @@ if(
             }
         }
     }
+    #. "$($PSScriptRoot)\ci\WikiUpdater.ps1"
 }